I change the structure of the code. #3
